Add dependency injection for params & Type Hints
You can now pass an optional container param to the `Router` constructor. The container must conform to PSR 11, allowing us to work with a range of DI containers.

Internally we use the `PHP-DI/Invoker` package to provide us with a generic `call()` function that makes it simple to resolve dependencies for classes and methods for any PSR 11 compatible container.

If you don’t provide a container in the constructor, the Router falls back to the previous behaviour and will pass a `RouteParams` to the closure/controller.